Just a comment on the short circuting code path for arrival SURBs
Also I emailed this to Christian :

This is mostly me thinking out loud.  I should add it to the .tex file,
but first I'll write it as an email :

There are two types of SURBs for unwinding :

Delivery SURBs are the ones that arrived at another node, like an
aggregation point.  I unwind them because there is a 16 byte PacketName
that node derives from the key exchange and writes to the SURB log.

Arrival SURBs arrive at the sender's machine.  I've been writing them to
use this PacketName too, but they differ in that they do not need to do
and immediately undo the last layer of encryption for the body and surb
log.

Arrival SURBs could avoid even this final public key operation if they
were stored according to the expected alpha.  Aside from being faster
for the client, this avoids the client even needing a routing key, which
simplifies things.  It cannot be easily modified to support encoding a
seed for the SURB inside the SURB itself however

Now the question becomes : How must do we care about supporting SURBs
generated by another device or party?  It's maybe not worth supporting
this code path where clients do not have routing keys, although clients
might not advertise their routing keys even if they need them.

Jeff